Update README.md
Flex sensors are included into a speaking system, and their resistance value changes in response 
to the user's set gestures. The Arduino Mega 2560 microcontroller processes the gesture data, and 
the matching speech output is displayed though LCD in text as well as played through the speaker 
in the specified languages.